requirements_clean.txt
streamlit
pandas
openpyxl
bcrypt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
streamlit-sortables
streamlit-card
plotly
fpdf2
